test(discord): split send behavior coverage
